scripts/requirements.txt
# E-Nose Python Dependencies
# Install with:  pip install -r requirements.txt

pyserial>=3.5
